Refactor page_size calculation; PEP fixes.
